Split all Resin-API-related functions into a separate package
