+++ provisioner/helpers.go
@@ -1,7 +1,6 @@
 package provisioner
 
 import (
-	"bytes"
 	"crypto/rand"
 	"fmt"
 	"io"
@@ -174,42 +173,6 @@ func setSupervisorTag() error {
 
 		return setEnvFileFields(SUPERVISOR_CONF_PATH, fields)
 	}
-}
-
-// Simple http GET request helper
-func getUrl(url string) ([]byte, int, error) {
-	resp, err := http.Get(url)
-	if err != nil {
-		return nil, 0, err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, 0, err
-	}
-	return body, resp.StatusCode, nil
-}
-
-// Simple http POST helper
-func postUrl(url string, bodyType string, body []byte) ([]byte, int, error) {
-	resp, err := http.Post(url, bodyType, bytes.NewBuffer(body))
-	if err != nil {
-		return nil, 0, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, 0, err
-	}
-	return b, resp.StatusCode, nil
-}
-
-func isHttpSuccess(status int) bool {
-	return status/100 == 2
-}
-
-func pineQueryEscape(s string) string {
-	return strings.Replace(s, " ", "%20", -1)
 }
 
 func randomHexString(byteLength uint32) (str string, err error) {
